refactor(admin): colocate premium feature forms
