+++ src/app/(admin)/admin/premium-features/_components/PremiumFeatureDialog.tsx
@@ -1,40 +1,118 @@
-"use client";
+import { Button, DialogFooter, Input, TypographyMuted, Textarea, Label } from "@/ui/components/atoms";
 
-import { useActionState, useEffect } from "react";
-import { toast } from "sonner";
 
-import type { PremiumFeature } from "@/services";
-import { updatePremiumFeature } from "@/actions";
+
+
+import type { PremiumFeature } from "@/core/domain";
+import { Alert } from "@/ui/components/molecules";
+import { TextField } from "@/ui/components/organisms";
 import type { APIResponse } from "@/core/domain";
-import { hasFieldErrors } from "@/core/utils";
-import { PremiumFeatureDialog as PurePremiumFeatureDialog } from "@/ui/components/organisms";
-export function PremiumFeatureDialog({
-  premiumFeature,
-}: {
-  premiumFeature: PremiumFeature;
-}) {
-  const [state, action, pending] = useActionState<
-    APIResponse<{ message: string }>,
-    FormData
-  >(updatePremiumFeature, null);
+import { getFieldError } from "@/core/utils";
 
-  useEffect(() => {
-    if (!state) return;
-    if (state.success === true) {
-      toast.success(state.data.message);
-    } else {
-      if (!hasFieldErrors(state.error)) {
-        toast.error(state.error.message);
-      }
-    }
-  }, [state]);
+interface PremiumFeatureDialogProps {
+  premiumFeature: PremiumFeature;
+  action: (formData: FormData) => void;
+  pending: boolean;
+  state: APIResponse<{ message: string }> | null;
+}
+
+export function PremiumFeatureDialog({
+  premiumFeature: feature,
+  action,
+  pending,
+  state,
+}: PremiumFeatureDialogProps) {
+  const codeError = getFieldError(state, "code");
+  const labelError = getFieldError(state, "label");
+  const descriptionError = getFieldError(state, "description");
+  const additionalPriceError = getFieldError(state, "additionalPrice");
 
   return (
-    <PurePremiumFeatureDialog
-      premiumFeature={premiumFeature}
-      action={action}
-      pending={pending}
-      state={state}
-    />
+    <form action={action}>
+      <div className="space-y-4 py-4">
+        <TextField
+            id="code"
+            name="code"
+            type="text"
+            placeholder="예: ANIMATION"
+            defaultValue={feature.code}
+            required
+            error={codeError}
+          >
+            기능 코드 *
+          </TextField>
+          <TypographyMuted>
+            영문 대문자와 언더스코어만 사용 가능합니다.
+          </TypographyMuted>
+
+        <TextField
+            id="label"
+            name="label"
+            type="text"
+            placeholder="예: 애니메이션 효과"
+            defaultValue={feature.label}
+            required
+            error={labelError}
+          >
+            기능 이름 *
+          </TextField>
+
+        <div className="space-y-2">
+          <Label htmlFor="description">
+            기능 설명 *
+          </Label>
+          <Textarea
+            id="description"
+            name="description"
+            placeholder="기능에 대한 자세한 설명을 입력하세요."
+            rows={3}
+            defaultValue={feature.description}
+            required
+          />
+          {descriptionError && <Alert type="error">{descriptionError}</Alert>}
+        </div>
+
+        <div className="space-y-2">
+          <Label htmlFor="additionalPrice">
+            추가 비용 *
+          </Label>
+          <div className="relative">
+            <Input
+              id="additionalPrice"
+              name="additionalPrice"
+              type="number"
+              placeholder="0"
+              min={0}
+              step={1000}
+              defaultValue={feature.additionalPrice}
+              required
+              className="pr-12"
+            />
+            <span className="text-muted-foreground absolute top-1/2 right-3 -translate-y-1/2 text-sm">
+              원
+            </span>
+          </div>
+          {additionalPriceError && (
+            <Alert type="error">{additionalPriceError}</Alert>
+          )}
+        </div>
+
+        <input
+          type="hidden"
+          id="featureId"
+          name="featureId"
+          value={feature._id}
+        />
+      </div>
+
+      <DialogFooter>
+        <Button type="button" variant="outline">
+          취소
+        </Button>
+        <Button type="submit" disabled={pending}>
+          {pending ? "수정 중..." : "수정"}
+        </Button>
+      </DialogFooter>
+    </form>
   );
 }
