+++ src/app/(admin)/admin/premium-features/_components/index.tsx
@@ -1,2 +1,1 @@
-export { PremiumFeatureDialog } from "./PremiumFeatureDialog";
 export { PremiumFeatureCardAction } from "./PremiumFeatureCardAction";
