+++ src/app/(admin)/admin/premium-features/new/_containers/PremiumFeatureRegistrationForm.tsx
@@ -1,0 +1,25 @@
+"use client";
+
+import { useActionState, useEffect } from "react";
+
+import { createPremiumFeature } from "@/actions";
+import type { APIResponse } from "@/core/domain";
+import { PremiumFeatureRegistrationForm as PurePremiumFeatureRegistrationForm } from "../_components/PremiumFeatureRegistrationForm";
+export function PremiumFeatureRegistrationForm() {
+  const [state, action, pending] = useActionState<
+    APIResponse<{ message: string }>,
+    FormData
+  >(createPremiumFeature, null);
+
+  useEffect(() => {
+    if (state && state.success && state.data) alert(state.data.message);
+  }, [state]);
+
+  return (
+    <PurePremiumFeatureRegistrationForm
+      action={action}
+      pending={pending}
+      state={state}
+    />
+  );
+}
