+++ src/app/(admin)/admin/premium-features/new/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { PremiumFeatureRegistrationForm } from "./PremiumFeatureRegistrationForm";
